Refactor: Change module name to CloneContentContext
